+++ lib/utils/image_filter_utils.dart
@@ -80,9 +80,9 @@ void brightness(Uint8List bytes, num adj) {
   adj = (adj < -1) ? -1 : adj;
   adj = ~~(255 * adj).round();
   for (int i = 0; i < bytes.length; i += 4) {
-    bytes[i] = clampPixel(bytes[i] + adj);
-    bytes[i + 1] = clampPixel(bytes[i + 1] + adj);
-    bytes[i + 2] = clampPixel(bytes[i + 2] + adj);
+    bytes[i] = clampPixel((bytes[i] + adj).toInt());
+    bytes[i + 1] = clampPixel((bytes[i + 1] + adj).toInt());
+    bytes[i + 2] = clampPixel((bytes[i + 2] + adj).toInt());
   }
 }
 
@@ -92,9 +92,9 @@ void hueSaturation(Uint8List bytes, num adj) {
     var hsv = rgbToHsv(bytes[i], bytes[i + 1], bytes[i + 2]);
     hsv[1] *= adj;
     var rgb = hsvToRgb(hsv[0], hsv[1], hsv[2]);
-    bytes[i] = clampPixel(rgb[0]);
-    bytes[i + 1] = clampPixel(rgb[1]);
-    bytes[i + 2] = clampPixel(rgb[2]);
+    bytes[i] = clampPixel(rgb[0].toInt());
+    bytes[i + 1] = clampPixel(rgb[1].toInt());
+    bytes[i + 2] = clampPixel(rgb[2].toInt());
   }
 }
